+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Moved mod auto-updating to CCModManager
 - Added CCModManager support
 - Added "Button denied" sound to the sound glossary
+- Added equipment menu support
 
 ## [0.5.7] 2024-02-06