+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added "Button denied" sound to the sound glossary
 - Added equipment menu support
 - Added quick menu item submenu support
+- Added analysis menu help menu
 
 ## [0.5.7] 2024-02-06